feat: oapi3 transform completed[]
