--- b.json
@@ -1,83 +1,0 @@
-{
-  "id": "0bad133f-1cfd-47c2-b33c-94279a413397",
-  "name": "Auth",
-  "description": "",
-  "type": "REQUEST",
-  "request": {
-    "method": "GET",
-    "url": "?queryParamKey=queryParamValue",
-    "body": {
-      "raw": "{\n\t\"jsonKey\": \"jsonValue\"\n}",
-      "urlencoded": [
-        {
-          "key": "",
-          "value": "",
-          "checked": false
-        }
-      ],
-      "formdata": {
-        "text": [
-          {
-            "key": "formDataTextKey",
-            "value": "formDataTextValue",
-            "checked": true
-          },
-          {
-            "key": "",
-            "value": "",
-            "checked": false
-          }
-        ],
-        "file": [
-          {
-            "key": "formDataFileKey",
-            "value": "/Users/nayanlakhwani/Downloads/giphy.gif",
-            "checked": true,
-            "base": "#@#/Users/nayanlakhwani/Downloads/giphy.gif"
-          },
-          {
-            "key": "",
-            "value": "",
-            "checked": false
-          }
-        ]
-      }
-    },
-    "headers": [
-      {
-        "key": "headerKey",
-        "value": "headerValue",
-        "checked": true
-      },
-      {
-        "key": "",
-        "value": "",
-        "checked": false
-      }
-    ],
-    "queryParams": [
-      {
-        "key": "queryParamKey",
-        "value": "queryParamValue",
-        "checked": true
-      },
-      {
-        "key": "",
-        "value": "",
-        "checked": false
-      }
-    ],
-    "auth": {
-      "bearerToken": "bearerToken",
-      "basicAuth": {
-        "username": "basicAuthUsername",
-        "password": "basicAuthPassword"
-      },
-      "apiKey": {
-        "authKey": "apiKey",
-        "authValue": "apikeyValue",
-        "addTo": "Header"
-      }
-    }
-  }
-}
