+++ src/modules/common/services/parser.service.ts
@@ -1,23 +1,18 @@
 import SwaggerParser from "@apidevtools/swagger-parser";
 import {
-  BodyModeEnum,
   Collection,
   CollectionItem,
   ItemTypeEnum,
-  RequestBody,
-  RequestMetaData,
   SourceTypeEnum,
 } from "../models/collection.model";
-import { OpenAPI303, ParameterObject } from "../models/openapi303.model";
-import { HTTPMethods } from "fastify";
+import { OpenAPI303 } from "../models/openapi303.model";
 import { Injectable } from "@nestjs/common";
 import { ContextService } from "./context.service";
-import { v4 as uuidv4 } from "uuid";
 import { CollectionService } from "@src/modules/workspace/services/collection.service";
 import { WithId } from "mongodb";
-import { resolveAllComponentRefs } from "./helper/parser.helper";
+import { resolveAllRefs } from "./helper/parser.helper";
 import { OpenAPI20 } from "../models/openapi20.model";
-import { createCollectionItems } from "./helper/oapi2.transformer";
+import { createCollectionItems } from "./helper/oapi.transformer";
 
 @Injectable()
 export class ParserService {
@@ -38,75 +33,15 @@ export class ParserService {
     let openApiDocument = (await SwaggerParser.parse(file)) as
       | OpenAPI303
       | OpenAPI20;
-    const baseUrl = this.getBaseUrl(openApiDocument);
     let existingCollection: WithId<Collection> | null = null;
     let folderObjMap = new Map();
     const user = await this.contextService.get("user");
     if (openApiDocument.hasOwnProperty("components")) {
-      openApiDocument = resolveAllComponentRefs(openApiDocument) as OpenAPI303;
-      for (const [key, value] of Object.entries(openApiDocument.paths)) {
-        //key will be endpoints /put and values will its apis post ,put etc
-        for (const [innerKey, innerValue] of Object.entries(value)) {
-          //key will be api methods and values will it's desc
-          const requestObj: CollectionItem = {} as CollectionItem;
-          requestObj.name = key;
-          requestObj.description = innerValue.description;
-          requestObj.type = ItemTypeEnum.REQUEST;
-          requestObj.source = SourceTypeEnum.SPEC;
-          requestObj.id = uuidv4();
-          requestObj.isDeleted = false;
-          requestObj.request = {} as RequestMetaData;
-          requestObj.request.method = innerKey.toUpperCase() as HTTPMethods;
-          requestObj.request.operationId = innerValue.operationId;
-          requestObj.request.url = baseUrl + key;
-          if (innerValue.parameters?.length) {
-            requestObj.request.queryParams = innerValue.parameters.filter(
-              (param: ParameterObject) => param.in === "query",
-            );
-            requestObj.request.pathParams = innerValue.parameters.filter(
-              (param: ParameterObject) => param.in === "path",
-            );
-            requestObj.request.headers = innerValue.parameters.filter(
-              (param: ParameterObject) => param.in === "header",
-            );
-          }
-          if (innerValue.requestBody) {
-            requestObj.request.body = [];
-            const bodyTypes = innerValue.requestBody.content;
-            for (const [type, schema] of Object.entries(bodyTypes)) {
-              const body: RequestBody = {} as RequestBody;
-              body.type = Object.values(BodyModeEnum).find(
-                (enumMember) => enumMember === type,
-              ) as BodyModeEnum;
-              body.schema = (schema as any).schema;
-              requestObj.request.body.push(body);
-            }
-          }
-          //Add to a folder
-          const tag = innerValue.tags ? innerValue.tags[0] : "default";
-          const tagArr =
-            openApiDocument?.tags?.length > 0 &&
-            openApiDocument.tags.filter((tagObj) => {
-              return tagObj.name === tag;
-            });
-          let folderObj: CollectionItem = folderObjMap.get(tag);
-          if (!folderObj) {
-            folderObj = {} as CollectionItem;
-            folderObj.name = tag;
-            folderObj.description = tagArr ? tagArr[0].description : "";
-            folderObj.isDeleted = false;
-            folderObj.type = ItemTypeEnum.FOLDER;
-            folderObj.id = uuidv4();
-            folderObj.items = [];
-          }
-          folderObj.items.push(requestObj);
-          folderObjMap.set(folderObj.name, folderObj);
-        }
-      }
+      openApiDocument = resolveAllRefs(openApiDocument) as OpenAPI303;
     } else if (openApiDocument.hasOwnProperty("definitions")) {
-      openApiDocument = resolveAllComponentRefs(openApiDocument) as OpenAPI20;
-      folderObjMap = createCollectionItems(openApiDocument, user);
+      openApiDocument = resolveAllRefs(openApiDocument) as OpenAPI20;
     }
+    folderObjMap = createCollectionItems(openApiDocument, user);
     const itemObject = Object.fromEntries(folderObjMap);
     let items: CollectionItem[] = [];
     let totalRequests = 0;
@@ -284,13 +219,5 @@ export class ParserService {
     });
 
     return mergedArray;
-  }
-  getBaseUrl(openApiDocument: any): string {
-    const basePath = openApiDocument.basePath ? openApiDocument.basePath : "";
-    if (openApiDocument.host) {
-      return "https://" + openApiDocument.host + basePath;
-    } else {
-      return "http://localhost:{{PORT}}" + basePath;
-    }
   }
 }
