--- test.js
@@ -1,2 +1,0 @@
-
-console.log(util.inspect(transformedPaths, false, null, true));
